requirements.txt
google-generativeai
python-dotenv
fastapi
uvicorn
pydantic
httpx
google-generativeai>=0.5.0
python-dotenv>=1.0.1
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
pydantic>=2.5.0
httpx>=0.25.0
